requirements.txt
asgiref==3.2.3
boto==2.49.0
boto3==1.11.13
botocore==1.14.13
certifi==2019.11.28
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
django-pagedown==2.0.3
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
